Refactor armor registries and naming
