+++ src/main/java/com/ultreon/mods/pixelguns/armor/ArmoredArmor.java
@@ -1,6 +1,7 @@
 package com.ultreon.mods.pixelguns.armor;
 
 import com.ultreon.mods.pixelguns.item.ModCreativeTab;
+import com.ultreon.mods.pixelguns.registry.ArmorRegistry;
 import net.fabricmc.fabric.api.item.v1.FabricItemSettings;
 import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.item.ArmorItem;
@@ -13,8 +14,8 @@ public class ArmoredArmor extends ArmorItem implements IAnimatable {
 	
 	private final AnimationFactory factory = GeckoLibUtil.createFactory(this);
 
-	public ArmoredArmor() {
-		super(ModArmorMaterials.ARMORED, EquipmentSlot.CHEST, new FabricItemSettings().group(ModCreativeTab.MISC));
+	public ArmoredArmor(EquipmentSlot equipmentSlot) {
+		super(ArmorRegistry.ARMORED, equipmentSlot, new FabricItemSettings().group(ModCreativeTab.MISC));
 	}
 
 	@Override
