+++ src/main/java/com/ultreon/mods/pixelguns/armor/HazardArmor.java
@@ -1,8 +1,10 @@
 package com.ultreon.mods.pixelguns.armor;
 
+import com.ultreon.mods.pixelguns.item.ModCreativeTab;
+import com.ultreon.mods.pixelguns.registry.ArmorRegistry;
+import net.fabricmc.fabric.api.item.v1.FabricItemSettings;
 import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.item.ArmorItem;
-import net.minecraft.item.ArmorMaterial;
 import software.bernie.geckolib3.core.IAnimatable;
 import software.bernie.geckolib3.core.manager.AnimationData;
 import software.bernie.geckolib3.core.manager.AnimationFactory;
@@ -12,8 +14,8 @@ public class HazardArmor extends ArmorItem implements IAnimatable {
 	
 	private final AnimationFactory factory = GeckoLibUtil.createFactory(this);
 
-	public HazardArmor(ArmorMaterial materialIn, EquipmentSlot slot, Settings builder) {
-		super(materialIn, slot, builder);
+	public HazardArmor(EquipmentSlot slot) {
+		super(ArmorRegistry.HAZARD, slot, new FabricItemSettings().group(ModCreativeTab.MISC));
 	}
 
 	@Override
